requirements.txt
Django==4.2.7
torch
torchvision
diffusers
transformers
accelerate
Pillow
safetensors
huggingface-hub